pyproject.toml
[project]
name = "qrypt"
version = "0.1.0"
description = "Simple Example CRUD API / UI for exploring Crypto Currencies"
readme = "README.md"
authors = [
    { name = "Chris Ward", email = "chris@calmrat.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.11.16",
    "alembic>=1.15.2",
    "asyncio>=3.4.3",
    "click>=8.1.3",
    "duckdb>=1.2.1",
    "fastapi[standard]>=0.115.0",
    "httpx>=0.28.1",
    "ipdb>=0.13.13",
    "mypy>=1.15.0",
    "pandas>=2.2.3",
    "pylance>=0.25.2",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.44.1",
    "uvicorn>=0.18.3",
]

[project.scripts]
init_db = "qrypt.core.ops.db:init_db"
drop_db = "qrypt.core.ops.db:drop_db"
pull_tokens = "qrypt.tokens.services.coingecko.ops.admin:pull_tokens"

[build-system]
requires = [
    "hatchling",
    # Package build dependencies
    "pip>=23.2.1",
    "poetry>=1.5.0",
    # Testing dependencies
    "pytest>=8.3.5",
    # validation dependencies
    "mypy>=1.5.0",
    "ruff>=0.0.263",
    "black>=23.9.1",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
    packages = ["src/qrypt"]

[tool.mypy]
ignore_missing_imports = true
python_version = "3.13"
